+++ index.js
@@ -1,66 +1,90 @@
 var fs = require('fs')
 var path = require('path')
-var quickTemp = require('quick-temp')
+var mergeTrees = require('broccoli-merge-trees')
+var Writer = require('broccoli-writer')
 var helpers = require('broccoli-kitchen-sink-helpers')
 
 
-function Tree (baseDirectory) {
-  this.baseDirectory = baseDirectory
-  this._mappings = []
+// Almost the same as using a plain string for a tree; but unlike a plain
+// string, Broccoli won't watch this
+function UnwatchedTree (dir) { this.dir = dir }
+UnwatchedTree.prototype.read = function (readTree) { return this.dir }
+UnwatchedTree.prototype.cleanup = function () { }
+
+
+
+MainFilePicker.prototype = Object.create(Writer.prototype)
+MainFilePicker.prototype.constructor = MainFilePicker
+function MainFilePicker (dir, mainFiles) {
+  this.dir = dir
+  this.mainFiles = mainFiles
 }
 
-Tree.prototype.map = function (src, dest) {
-  helpers.assertAbsolutePaths([dest])
-  this._mappings.push([src, dest])
-  return this
-}
-
-Tree.prototype.read = function (readTree) {
-  quickTemp.makeOrRemake(this, '_tmpDestDir')
-
-  if (this._mappings.length === 0) throw new Error('No tree map for ' + this)
-
-  for (var i = 0; i < this._mappings.length; i++) {
-    var src = this._mappings[i][0], dest = this._mappings[i][1]
-    helpers.copyRecursivelySync(
-      path.join(this.baseDirectory, src),
-      path.join(this._tmpDestDir, dest))
+MainFilePicker.prototype.write = function (readTree, destDir) {
+  for (var i = 0; i < this.mainFiles.length; i++) {
+    var mainFile = this.mainFiles[i]
+    try {
+      if (mainFile.indexOf('*') !== -1 || mainFile.indexOf('?') !== -1) {
+        throw new Error('In ' + this.dir + ': Glob patterns in bower\'s `main` property are not allowed: ' + mainFile)
+      }
+      helpers.copyPreserveSync(
+        path.join(this.dir, mainFile),
+        path.join(destDir, path.basename(mainFile)))
+    } catch (err) {
+      if (err.code !== 'ENOENT') throw err
+      throw new Error('In bower directory ' + this.dir + ': Cannot find file specified in bower `main` property: ' + mainFile)
+    }
   }
-
-  return this._tmpDestDir
 }
 
-Tree.prototype.cleanup = function () {
-  quickTemp.remove(this, '_tmpDestDir')
+
+module.exports = findBowerTrees
+function findBowerTrees () {
+  var bowerDir = require('bower-config').read().directory // note: this relies on cwd
+  if (bowerDir == null) throw new Error('Bower did not return a directory')
+  var entries = fs.readdirSync(bowerDir)
+  var directories = entries.filter(function (f) {
+    return fs.statSync(path.join(bowerDir, f)).isDirectory()
+  })
+  return directories.map(function (dir) {
+    return treeFromDirectory(path.join(bowerDir, dir))
+  })
 }
 
-Tree._fromDirectory = function (dir) {
+
+var treeFromDirectory = function (dir) {
   // Guess some reasonable defaults
   var options = bowerOptionsForDirectory(dir)
   var treeOptions = {}
   if (options.main != null) {
     var main = options.main
     if (typeof main === 'string') main = [main]
-    if (!Array.isArray(main)) throw new Error(dir + ': Expected "main" bower option to be array or string')
+    if (!Array.isArray(main)) throw new Error(dir + ': Expected `main` bower option to be array or string')
     treeOptions.main = main
   }
 
-  var tree = new Tree(dir)
-  // It's not clear that this is a sensible heuristic to hard-code here
+  // We may need to use multiple auxiliary tree objects to represent this
+  // bower tree
+  var trees = []
+
+  // Terrible heuristics follow
   if (fs.existsSync(path.join(dir, 'lib'))) {
-    tree.map('lib', '/')
+    trees.push(new UnwatchedTree(path.join(dir, 'lib')))
   }
   if (treeOptions.main) {
-    for (var i = 0; i < treeOptions.main.length; i++) {
-      tree.map(treeOptions.main[i], '/' + path.basename(treeOptions.main[i]))
-    }
-  } else {
-    // Map repo root into namespace root. We should perhaps avoid this, as
-    // it causes clutter.
-    tree.map('', '/')
+    trees.push(new MainFilePicker(dir, treeOptions.main))
+  }
+  if (trees.length === 0) {
+    // Map repo root into namespace root. This is almost always wrong, but we
+    // have it as a stop-gap.
+    trees.push(new UnwatchedTree(dir))
   }
 
-  return [tree]
+  if (trees.length === 1) {
+    return trees[0]
+  } else {
+    return mergeTrees(trees, { overwrite: true })
+  }
 
   function bowerOptionsForDirectory (dir) {
     var options = {}
@@ -81,29 +105,4 @@ Tree._fromDirectory = function (dir) {
     })
     return options
   }
-}
-
-
-// This is exposed as API, but needs refactoring and will likely change
-module.exports = findBowerTrees
-function findBowerTrees () {
-  var bowerDir = require('bower-config').read().directory // note: this relies on cwd
-  if (bowerDir == null) throw new Error('Bower did not return a directory')
-  var entries = fs.readdirSync(bowerDir)
-  var directories = entries.filter(function (f) {
-    return fs.statSync(path.join(bowerDir, f)).isDirectory()
-  })
-  var files = entries.filter(function (f) {
-    var stat = fs.statSync(path.join(bowerDir, f))
-    return stat.isFile() || stat.isSymbolicLink()
-  })
-  var trees = []
-  for (var i = 0; i < directories.length; i++) {
-    trees = trees.concat(Tree._fromDirectory(path.join(bowerDir, directories[i])))
-  }
-  // Pick up files as well; this is for compatibility with EAK's vendor/loader.js
-  for (i = 0; i < files.length; i++) {
-    trees.push(new Tree(bowerDir).map('/' + files[i], '/' + files[i]))
-  }
-  return trees
 }